refactor: Move model service into models context directory
