+++ frontend/src/app/projects/models/choose-init/choose-init.component.ts
@@ -4,7 +4,7 @@
  */
 
 import { Component, EventEmitter, Output } from '@angular/core';
-import { ModelService } from 'src/app/services/model/model.service';
+import { ModelService } from 'src/app/projects/models/service/model.service';
 import { ProjectService } from 'src/app/services/project/project.service';
 
 @Component({
