+++ frontend/src/app/projects/models/init-model/init-model.component.ts
@@ -13,7 +13,10 @@ import {
 } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { combineLatest, filter, map, Subscription, switchMap, tap } from 'rxjs';
-import { Model, ModelService } from 'src/app/services/model/model.service';
+import {
+  Model,
+  ModelService,
+} from 'src/app/projects/models/service/model.service';
 import { ProjectService } from 'src/app/services/project/project.service';
 import {
   Tool,
