+++ frontend/src/app/projects/models/model-detail/model-detail.component.ts
@@ -9,11 +9,11 @@ import {
   T4CModel,
   T4CModelService,
 } from 'src/app/projects/models/model-source/t4c/service/t4c-model.service';
+import { ModelService } from 'src/app/projects/models/service/model.service';
 import {
   GetGitModel,
   GitModelService,
 } from 'src/app/projects/project-detail/model-overview/model-detail/git-model.service';
-import { ModelService } from 'src/app/services/model/model.service';
 import { ProjectService } from 'src/app/services/project/project.service';
 
 @Component({
