+++ frontend/src/app/projects/models/model-source/choose-source.component.ts
@@ -4,7 +4,7 @@
  */
 
 import { Component, EventEmitter, Output } from '@angular/core';
-import { ModelService } from 'src/app/services/model/model.service';
+import { ModelService } from 'src/app/projects/models/service/model.service';
 import { ProjectService } from 'src/app/services/project/project.service';
 import { UserService } from 'src/app/services/user/user.service';
