+++ frontend/src/app/projects/models/model-source/t4c/manage-t4c-model/manage-t4c-model.component.ts
@@ -20,7 +20,7 @@ import {
   T4CModel,
   T4CModelService,
 } from 'src/app/projects/models/model-source/t4c/service/t4c-model.service';
-import { ModelService } from 'src/app/services/model/model.service';
+import { ModelService } from 'src/app/projects/models/service/model.service';
 import { ProjectService } from 'src/app/services/project/project.service';
 import {
   T4CInstance,
