+++ frontend/src/app/projects/project-wrapper/project-wrapper.component.ts
@@ -6,8 +6,11 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Subject, Subscription, connectable, map, switchMap } from 'rxjs';
+import {
+  Model,
+  ModelService,
+} from 'src/app/projects/models/service/model.service';
 import { ProjectUserService } from 'src/app/projects/project-detail/project-users/service/project-user.service';
-import { Model, ModelService } from 'src/app/services/model/model.service';
 import {
   Project,
   ProjectService,
